feat(ttf): add sub font
