+++ packages/core/src/2d/text/Font.ts
@@ -1,9 +1,8 @@
 import { RefObject } from "../../asset/RefObject";
 import { Engine } from "../../Engine";
-import { Texture2D } from "../../texture";
-import { CharInfo } from "./CharInfo";
-import { FontAtlas } from "../atlas/FontAtlas";
 import { TextUtils } from "./TextUtils";
+import { SubFont } from "./SubFont";
+import { FontStyle } from "../enums/FontStyle";
 
 /**
  * Font.
@@ -50,22 +49,8 @@ export class Font extends RefObject {
     }
   }
 
-  /** @internal */
-  static _createFromOS(engine: Engine, name: string = ""): Font {
-    const fontMap = Font._fontMap;
-    let font = fontMap[name];
-    if (font) {
-      return font;
-    }
-    font = new Font(engine, name);
-    font._addRefCount(1);
-    fontMap[name] = font;
-    return font;
-  }
-
   private _name: string = "";
-  private _fontAtlases: FontAtlas[] = [];
-  private _lastIndex: number = -1;
+  private _subFontMap: Record<string, SubFont> = {};
 
   /**
    * The name of the font object.
@@ -74,91 +59,33 @@ export class Font extends RefObject {
     return this._name;
   }
 
-  private constructor(engine: Engine, name: string = "") {
+  constructor(engine: Engine, name: string = "") {
     super(engine);
     this._name = name;
   }
 
-  /**
-   * @internal
-   */
-  _uploadCharTexture(charInfo: CharInfo): void {
-    const fontAtlases = this._fontAtlases;
-    let lastIndex = this._lastIndex;
-    if (lastIndex === -1) {
-      this._createFontAtlas();
-      lastIndex++;
+  /** @internal */
+  _getSubFont(fontSize: number, fontStyle: FontStyle): SubFont {
+    const key = `${fontSize}-${fontStyle}`;
+    const subFontMap = this._subFontMap;
+    let subFont = subFontMap[key];
+    if (subFont) {
+      return subFont;
     }
-    let fontAtlas = fontAtlases[lastIndex];
-    if (!fontAtlas.uploadCharTexture(charInfo)) {
-      fontAtlas = this._createFontAtlas();
-      fontAtlas.uploadCharTexture(charInfo);
-      lastIndex++;
-    }
-    this._lastIndex = lastIndex;
-    charInfo.data = null;
-  }
-
-  /**
-   * @internal
-   */
-  _addCharInfo(char: string, charInfo: CharInfo) {
-    const lastIndex = this._lastIndex;
-    charInfo.index = lastIndex;
-    this._fontAtlases[lastIndex].addCharInfo(char, charInfo);
-  }
-
-  /**
-   * @internal
-   */
-  _getCharInfo(char: string): CharInfo {
-    const fontAtlases = this._fontAtlases;
-    for (let i = 0, n = fontAtlases.length; i < n; ++i) {
-      const fontAtlas = fontAtlases[i];
-      const charInfo = fontAtlas.getCharInfo(char);
-      if (charInfo) {
-        return charInfo;
-      }
-    }
-    return null;
-  }
-
-  /**
-   * @internal
-   */
-  _getTextureByIndex(index: number): Texture2D {
-    const fontAtlas = this._fontAtlases[index];
-    if (fontAtlas) {
-      return fontAtlas.texture;
-    }
-    return null;
-  }
-
-  /**
-   * @internal
-   */
-  _getLastIndex(): number {
-    return this._lastIndex;
+    subFont = new SubFont(this.engine);
+    subFontMap[key] = subFont;
+    return subFont;
   }
 
   /**
    * @override
    */
   _onDestroy(): void {
-    const fontAtlases = this._fontAtlases;
-    for (let i = 0, n = fontAtlases.length; i < n; ++i) {
-      fontAtlases[i].destroy(true);
+    const styleFonts = Object.values(this._subFontMap);
+    for (let i = 0, l = styleFonts.length; i < l; ++i) {
+      styleFonts[i].destroy();
     }
-    fontAtlases.length = 0;
+    this._subFontMap = null;
     delete Font._fontMap[this._name];
-  }
-
-  private _createFontAtlas(): FontAtlas {
-    const { engine } = this;
-    const fontAtlas = new FontAtlas(engine);
-    const texture = new Texture2D(engine, 256, 256);
-    fontAtlas.texture = texture;
-    this._fontAtlases.push(fontAtlas);
-    return fontAtlas;
   }
 }
