+++ packages/core/src/2d/text/TextUtils.ts
@@ -94,7 +94,7 @@ export class TextUtils {
     const { fontSize, fontStyle } = renderer;
     const { name } = renderer.font;
     const fontString = TextUtils.getNativeFontString(name, fontSize, fontStyle);
-    const charFont = renderer._styleFont;
+    const charFont = renderer._subFont;
     const fontSizeInfo = TextUtils.measureFont(fontString);
     const subTexts = renderer.text.split(/(?:\r\n|\r|\n)/);
     const lines = new Array<string>();
@@ -179,7 +179,7 @@ export class TextUtils {
     const { fontSize, fontStyle } = renderer;
     const { name } = renderer.font;
     const fontString = TextUtils.getNativeFontString(name, fontSize, fontStyle);
-    const charFont = renderer._styleFont;
+    const charFont = renderer._subFont;
     const fontSizeInfo = TextUtils.measureFont(fontString);
     const lines = renderer.text.split(/(?:\r\n|\r|\n)/);
     const lineCount = lines.length;
